Extract common logic for HTTP exporters
