+++ exporters/otlp-http/trace/src/main/java/io/opentelemetry/exporter/otlp/http/trace/OtlpHttpSpanExporter.java
@@ -5,164 +5,31 @@
 
 package io.opentelemetry.exporter.otlp.http.trace;
 
-import io.opentelemetry.exporter.otlp.internal.ExporterMetrics;
-import io.opentelemetry.exporter.otlp.internal.ProtoRequestBody;
-import io.opentelemetry.exporter.otlp.internal.grpc.GrpcStatusUtil;
+import io.opentelemetry.exporter.otlp.internal.okhttp.OkHttpExporter;
 import io.opentelemetry.exporter.otlp.internal.traces.TraceRequestMarshaler;
 import io.opentelemetry.sdk.common.CompletableResultCode;
-import io.opentelemetry.sdk.internal.ThrottlingLogger;
 import io.opentelemetry.sdk.trace.data.SpanData;
 import io.opentelemetry.sdk.trace.export.SpanExporter;
-import java.io.IOException;
 import java.util.Collection;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
-import okhttp3.Call;
-import okhttp3.Callback;
-import okhttp3.Headers;
-import okhttp3.MediaType;
-import okhttp3.OkHttpClient;
-import okhttp3.Request;
-import okhttp3.RequestBody;
-import okhttp3.Response;
-import okhttp3.ResponseBody;
-import okio.BufferedSink;
-import okio.GzipSink;
-import okio.Okio;
 
 /** Exports spans using OTLP via HTTP, using OpenTelemetry's protobuf model. */
 @ThreadSafe
 public final class OtlpHttpSpanExporter implements SpanExporter {
 
-  private static final Logger internalLogger =
-      Logger.getLogger(OtlpHttpSpanExporter.class.getName());
+  private final OkHttpExporter<TraceRequestMarshaler> delegate;
 
-  private final ThrottlingLogger logger = new ThrottlingLogger(internalLogger);
-
-  private final ExporterMetrics exporterMetrics;
-  private final OkHttpClient client;
-  private final String endpoint;
-  @Nullable private final Headers headers;
-  private final boolean compressionEnabled;
-
-  OtlpHttpSpanExporter(
-      OkHttpClient client, String endpoint, @Nullable Headers headers, boolean compressionEnabled) {
-    this.exporterMetrics = ExporterMetrics.createHttpProtobuf("span");
-    this.client = client;
-    this.endpoint = endpoint;
-    this.headers = headers;
-    this.compressionEnabled = compressionEnabled;
+  OtlpHttpSpanExporter(OkHttpExporter<TraceRequestMarshaler> delegate) {
+    this.delegate = delegate;
   }
 
   /**
-   * Submits all the given spans in a single batch to the OpenTelemetry collector.
+   * Returns a new {@link OtlpHttpSpanExporter} using the default values.
    *
-   * @param spans the list of sampled Spans to be exported.
-   * @return the result of the operation
+   * @return a new {@link OtlpHttpSpanExporter} instance.
    */
-  @Override
-  public CompletableResultCode export(Collection<SpanData> spans) {
-    exporterMetrics.addSeen(spans.size());
-
-    TraceRequestMarshaler exportRequest = TraceRequestMarshaler.create(spans);
-
-    Request.Builder requestBuilder = new Request.Builder().url(endpoint);
-    if (headers != null) {
-      requestBuilder.headers(headers);
-    }
-    RequestBody requestBody = new ProtoRequestBody(exportRequest);
-    if (compressionEnabled) {
-      requestBuilder.addHeader("Content-Encoding", "gzip");
-      requestBuilder.post(gzipRequestBody(requestBody));
-    } else {
-      requestBuilder.post(requestBody);
-    }
-
-    CompletableResultCode result = new CompletableResultCode();
-
-    client
-        .newCall(requestBuilder.build())
-        .enqueue(
-            new Callback() {
-              @Override
-              public void onFailure(Call call, IOException e) {
-                exporterMetrics.addFailed(spans.size());
-                logger.log(
-                    Level.SEVERE,
-                    "Failed to export spans. The request could not be executed. Full error message: "
-                        + e.getMessage());
-                result.fail();
-              }
-
-              @Override
-              public void onResponse(Call call, Response response) {
-                if (response.isSuccessful()) {
-                  exporterMetrics.addSuccess(spans.size());
-                  result.succeed();
-                  return;
-                }
-
-                exporterMetrics.addFailed(spans.size());
-                int code = response.code();
-
-                String status = extractErrorStatus(response);
-
-                logger.log(
-                    Level.WARNING,
-                    "Failed to export spans. Server responded with HTTP status code "
-                        + code
-                        + ". Error message: "
-                        + status);
-                result.fail();
-              }
-            });
-
-    return result;
-  }
-
-  private static RequestBody gzipRequestBody(RequestBody requestBody) {
-    return new RequestBody() {
-      @Override
-      public MediaType contentType() {
-        return requestBody.contentType();
-      }
-
-      @Override
-      public long contentLength() {
-        return -1;
-      }
-
-      @Override
-      public void writeTo(BufferedSink bufferedSink) throws IOException {
-        BufferedSink gzipSink = Okio.buffer(new GzipSink(bufferedSink));
-        requestBody.writeTo(gzipSink);
-        gzipSink.close();
-      }
-    };
-  }
-
-  private static String extractErrorStatus(Response response) {
-    ResponseBody responseBody = response.body();
-    if (responseBody == null) {
-      return "Response body missing, HTTP status message: " + response.message();
-    }
-    try {
-      return GrpcStatusUtil.getStatusMessage(responseBody.bytes());
-    } catch (IOException e) {
-      return "Unable to parse response body, HTTP status message: " + response.message();
-    }
-  }
-
-  /**
-   * The OTLP exporter does not batch spans, so this method will immediately return with success.
-   *
-   * @return always Success
-   */
-  @Override
-  public CompletableResultCode flush() {
-    return CompletableResultCode.ofSuccess();
+  public static OtlpHttpSpanExporter getDefault() {
+    return builder().build();
   }
 
   /**
@@ -175,20 +42,30 @@ public final class OtlpHttpSpanExporter implements SpanExporter {
   }
 
   /**
-   * Returns a new {@link OtlpHttpSpanExporter} using the default values.
+   * Submits all the given spans in a single batch to the OpenTelemetry collector.
    *
-   * @return a new {@link OtlpHttpSpanExporter} instance.
+   * @param spans the list of sampled Spans to be exported.
+   * @return the result of the operation
    */
-  public static OtlpHttpSpanExporter getDefault() {
-    return builder().build();
+  @Override
+  public CompletableResultCode export(Collection<SpanData> spans) {
+    TraceRequestMarshaler exportRequest = TraceRequestMarshaler.create(spans);
+    return delegate.export(exportRequest, spans.size());
+  }
+
+  /**
+   * The OTLP exporter does not batch spans, so this method will immediately return with success.
+   *
+   * @return always Success
+   */
+  @Override
+  public CompletableResultCode flush() {
+    return CompletableResultCode.ofSuccess();
   }
 
   /** Shutdown the exporter. */
   @Override
   public CompletableResultCode shutdown() {
-    final CompletableResultCode result = CompletableResultCode.ofSuccess();
-    client.dispatcher().cancelAll();
-    exporterMetrics.unbind();
-    return result;
+    return delegate.shutdown();
   }
 }
