+++ exporters/otlp-http/trace/src/main/java/io/opentelemetry/exporter/otlp/http/trace/OtlpHttpSpanExporterBuilder.java
@@ -8,43 +8,36 @@ package io.opentelemetry.exporter.otlp.http.trace;
 import static io.opentelemetry.api.internal.Utils.checkArgument;
 import static java.util.Objects.requireNonNull;
 
-import io.opentelemetry.exporter.otlp.internal.TlsUtil;
-import java.net.URI;
-import java.net.URISyntaxException;
+import io.opentelemetry.exporter.otlp.internal.okhttp.OkHttpExporterBuilder;
+import io.opentelemetry.exporter.otlp.internal.traces.TraceRequestMarshaler;
 import java.time.Duration;
 import java.util.concurrent.TimeUnit;
-import javax.annotation.Nullable;
-import javax.net.ssl.SSLException;
-import javax.net.ssl.X509TrustManager;
-import okhttp3.Headers;
-import okhttp3.OkHttpClient;
 
 /** Builder utility for {@link OtlpHttpSpanExporter}. */
 public final class OtlpHttpSpanExporterBuilder {
 
-  private static final long DEFAULT_TIMEOUT_SECS = 10;
   private static final String DEFAULT_ENDPOINT = "http://localhost:4318/v1/traces";
 
-  private long timeoutNanos = TimeUnit.SECONDS.toNanos(DEFAULT_TIMEOUT_SECS);
-  private String endpoint = DEFAULT_ENDPOINT;
-  private boolean compressionEnabled = false;
-  @Nullable private Headers.Builder headersBuilder;
-  @Nullable private byte[] trustedCertificatesPem;
+  private final OkHttpExporterBuilder<TraceRequestMarshaler> delegate;
+
+  OtlpHttpSpanExporterBuilder() {
+    delegate = new OkHttpExporterBuilder<>("span", DEFAULT_ENDPOINT);
+  }
 
   /**
    * Sets the maximum time to wait for the collector to process an exported batch of spans. If
-   * unset, defaults to {@value DEFAULT_TIMEOUT_SECS}s.
+   * unset, defaults to {@value OkHttpExporterBuilder#DEFAULT_TIMEOUT_SECS}s.
    */
   public OtlpHttpSpanExporterBuilder setTimeout(long timeout, TimeUnit unit) {
     requireNonNull(unit, "unit");
     checkArgument(timeout >= 0, "timeout must be non-negative");
-    timeoutNanos = unit.toNanos(timeout);
+    delegate.setTimeout(timeout, unit);
     return this;
   }
 
   /**
    * Sets the maximum time to wait for the collector to process an exported batch of spans. If
-   * unset, defaults to {@value DEFAULT_TIMEOUT_SECS}s.
+   * unset, defaults to {@value OkHttpExporterBuilder#DEFAULT_TIMEOUT_SECS}s.
    */
   public OtlpHttpSpanExporterBuilder setTimeout(Duration timeout) {
     requireNonNull(timeout, "timeout");
@@ -57,21 +50,7 @@ public final class OtlpHttpSpanExporterBuilder {
    */
   public OtlpHttpSpanExporterBuilder setEndpoint(String endpoint) {
     requireNonNull(endpoint, "endpoint");
-
-    URI uri;
-    try {
-      uri = new URI(endpoint);
-    } catch (URISyntaxException e) {
-      throw new IllegalArgumentException("Invalid endpoint, must be a URL: " + endpoint, e);
-    }
-
-    if (uri.getScheme() == null
-        || (!uri.getScheme().equals("http") && !uri.getScheme().equals("https"))) {
-      throw new IllegalArgumentException(
-          "Invalid endpoint, must start with http:// or https://: " + uri);
-    }
-
-    this.endpoint = endpoint;
+    delegate.setEndpoint(endpoint);
     return this;
   }
 
@@ -84,18 +63,13 @@ public final class OtlpHttpSpanExporterBuilder {
     checkArgument(
         compressionMethod.equals("gzip") || compressionMethod.equals("none"),
         "Unsupported compression method. Supported compression methods include: gzip, none.");
-    if (compressionMethod.equals("gzip")) {
-      this.compressionEnabled = true;
-    }
+    delegate.setCompression(compressionMethod);
     return this;
   }
 
   /** Add header to requests. */
   public OtlpHttpSpanExporterBuilder addHeader(String key, String value) {
-    if (headersBuilder == null) {
-      headersBuilder = new Headers.Builder();
-    }
-    headersBuilder.add(key, value);
+    delegate.addHeader(key, value);
     return this;
   }
 
@@ -105,7 +79,7 @@ public final class OtlpHttpSpanExporterBuilder {
    * use the system default trusted certificates.
    */
   public OtlpHttpSpanExporterBuilder setTrustedCertificates(byte[] trustedCertificatesPem) {
-    this.trustedCertificatesPem = trustedCertificatesPem;
+    delegate.setTrustedCertificates(trustedCertificatesPem);
     return this;
   }
 
@@ -115,24 +89,6 @@ public final class OtlpHttpSpanExporterBuilder {
    * @return a new exporter's instance
    */
   public OtlpHttpSpanExporter build() {
-    OkHttpClient.Builder clientBuilder =
-        new OkHttpClient.Builder().callTimeout(Duration.ofNanos(timeoutNanos));
-
-    if (trustedCertificatesPem != null) {
-      try {
-        X509TrustManager trustManager = TlsUtil.trustManager(trustedCertificatesPem);
-        clientBuilder.sslSocketFactory(TlsUtil.sslSocketFactory(trustManager), trustManager);
-      } catch (SSLException e) {
-        throw new IllegalStateException(
-            "Could not set trusted certificate for OTLP HTTP connection, are they valid X.509 in PEM format?",
-            e);
-      }
-    }
-
-    Headers headers = headersBuilder == null ? null : headersBuilder.build();
-
-    return new OtlpHttpSpanExporter(clientBuilder.build(), endpoint, headers, compressionEnabled);
+    return new OtlpHttpSpanExporter(delegate.build());
   }
-
-  OtlpHttpSpanExporterBuilder() {}
 }
