+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/otlp/internal/okhttp/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** Utilities for HTTP exporters. */
+@ParametersAreNonnullByDefault
+package io.opentelemetry.exporter.otlp.internal.okhttp;
+
+import javax.annotation.ParametersAreNonnullByDefault;
